poster_generation/requirements.txt
langgraph>=0.2.45
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0
docling>=2.7.1
docling-core>=2.3.0
python-pptx>=1.0.2
Pillow>=10.0.0
opencv-python>=4.8.0
PyMuPDF>=1.24.0
pypdf>=5.0.0
pdfminer.six==20231228
pymupdf4llm>=0.0.17
requests>=2.32.0
beautifulsoup4>=4.12.0
urllib3>=2.0.0
torch==2.7.1
torchvision==0.22.1
transformers==4.52.4
diffusers>=0.25.0
accelerate>=0.30.0
easyocr>=1.7.0
numpy>=1.24.0
pandas>=2.0.0
datasets>=3.0.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
Jinja2>=3.1.0
pydantic>=2.0.0
tenacity>=8.0.0
typing-extensions>=4.8.0
tqdm>=4.66.0
imageio>=2.34.0
ffmpeg-python>=0.2.0
jupyter>=1.1.0
matplotlib>=3.7.0
seaborn>=0.12.0
aiohttp>=3.9.0
aiofiles>=24.0.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
jsonschema>=4.20.0
pyyaml>=6.0.0
markdown>=3.5.0
json_repair==0.35.0
git+https://github.com/Hadlay-Zhang/marker.git
pypandoc
qrcode[pil]
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6